Refactor: change project structure to use a trivia directory
