+++ trivia/routes/auth.py
@@ -1,9 +1,8 @@
 from flask import Blueprint, render_template, redirect, url_for, request, flash
 from flask_login import login_user
-from sqlalchemy.sql import functions as func
 from werkzeug.security import generate_password_hash, check_password_hash
-from db import db
-from models import Question, User
+from trivia.db import db
+from trivia.models import User
 
 auth_bp = Blueprint("auth", __name__)
