+++ trivia/static/style.css
