+++ trivia/templates/register.html
